Move QuickLoginViewModel to its own file
